Fix Bug #213
Because of the way the page was loaded inside the tab, binding was
happening before the ViewModel had been initialized so the ToggleSwitch
was first set wrong and then reset after the command was assigned. This
got the ToggleSwitch setting backwards and in turn caused the wrong
theme to be used.