Add ability to remove code from original file
Extended PostAction merging to be able to remove something from the
original file if so desired.

This commit also includes modifications to features to remove the
unneeded `yield break;` in ActivationService.GetActivationHandlers -
#252

Only does C# as that's all that's needed at the moment. - Can't yet
imagine a sceanrio where we'd want to remove XAML.