Webview CB Template conditionally displays acrylic
The command bar on the bottom of the webview overlays and displays acrylic in accordance with design principles on RS3. In RS2 they default behavior from previous releases occurs. Aditionally, webview content is shifted a number of line breaks up to accomodate overlapping acrylic bar without cutting off the bottom of the page